Create reusable code changes table
